make scraping a class with per-site implementations
