+++ api/src/service/scrape.ts
@@ -1,332 +1,25 @@
-import axios from "axios";
-import cheerio from "cheerio";
 import { Moment } from "moment-timezone";
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-import unescape from "unescape";
-import https from "https";
 import {
   Comic,
-  ComicFailureMode,
   failureModes,
   IComic,
   ISyndication,
   Syndication,
 } from "../db-models/comic-syndication";
 import { ScrapeAndSaveAllComicsOptions } from "../api-models/scrape-options";
-
-const sites = {
-  gocomics: {
-    name: "GOCOMICS",
-    description: "GoComics",
-    id: 0,
-  },
-  comicskingdom: {
-    name: "COMICS_KINGDOM",
-    description: "Comics Kingdom",
-    id: 1,
-  },
-  arcamax: {
-    name: "ARCAMAX",
-    description: "ArcaMax",
-    id: 2,
-  },
-  tundra: {
-    name: "Tundra",
-    description: "Tundra",
-    id: 3,
-  },
-  xkcd: {
-    name: "XKCD",
-    description: "xkcd",
-    id: 4,
-  },
-  phdcomics: {
-    name: "PHDCOMICS",
-    description: "PhD Comics",
-    id: 5,
-  },
-  explosm: {
-    name: "EXPLOSM",
-    description: "Explosm",
-    id: 6,
-  },
-  arcamaxeditorials: {
-    name: "ARCAMAX_EDITORIALS",
-    description: "ArcaMax Editorials",
-    id: 7,
-  },
-  dinosaur: {
-    name: "DINOSAUR",
-    description: "Dinosaur",
-    id: 8,
-  },
-  // TODO(ecarrel): delete SMBC? It's unnecessary since we already get it from gocomics.
-  smbc: {
-    name: "SMBC",
-    description: "Saturday Morning Breakfast Cereal",
-    id: 9,
-  },
-  threewordphrase: {
-    name: "THREE_WORD_PHRASE",
-    description: "Three Word Phrase",
-    id: 10,
-  },
-};
-
-interface ScrapeResult {
-  success: boolean;
-  imageUrl: string | null;
-  failureMode: ComicFailureMode | null;
-  imageCaption: string | null;
-}
-
-const scrapeFailure = (failureMode: ComicFailureMode): ScrapeResult => ({
-  success: false,
-  imageUrl: null,
-  failureMode,
-  imageCaption: null,
-});
-
-const scrapeSuccess = (
-  imageUrl: string | undefined,
-  imageCaption?: string | null,
-): ScrapeResult => {
-  if (imageUrl == null || imageUrl.length === 0) {
-    return scrapeFailure(failureModes.UNKNOWN);
-  }
-  return {
-    success: true,
-    imageUrl,
-    failureMode: null,
-    imageCaption: imageCaption || null,
-  };
-};
-
-const cheerioRequest = async (url: string) => {
-  try {
-    const response = await axios({
-      method: "GET",
-      url,
-      responseType: "text",
-      httpsAgent: new https.Agent({ rejectUnauthorized: false }),
-    });
-    // TODO(ecarrel): check if response.request.res.responseUrl !== url to detect redirects.
-    return cheerio.load(response.data);
-  } catch (err) {
-    return null;
-  }
-};
-
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-// eslint-disable-next-line no-underscore-dangle,@typescript-eslint/no-unused-vars
-const _producesNonEmptyResponse = async (url: string) => {
-  const response = await axios({
-    method: "GET",
-    url,
-    responseType: "stream",
-  }).catch(() => false);
-  if (typeof response === "boolean") {
-    return false;
-  }
-  const content = response.data.read(1);
-  return content != null;
-};
-
-const producesSuccessResponse = async (url: string) => {
-  const response = await axios({
-    method: "GET",
-    url,
-    responseType: "stream",
-  }).catch(() => false);
-  if (typeof response === "boolean") {
-    return false;
-  }
-  return response.status === 200;
-};
+import { sites } from "./scraper/sites";
+import { scrapeFailure, ScrapeResult } from "./scraper/scraper";
 
 export const scrapeComicForSyndication = async (
   syndication: ISyndication,
   date: Moment,
 ): Promise<ScrapeResult> => {
-  const { site_id: siteId, theiridentifier: theirIdentifier } = syndication;
-  switch (siteId) {
-    case sites.gocomics.id: {
-      const url = `http://www.gocomics.com/${theirIdentifier}/${date.format(
-        "YYYY/MM/DD",
-      )}`;
-      const $ = await cheerioRequest(url);
-      if ($ === null) {
-        return scrapeFailure(failureModes.GOCOMICS_REJECTION);
-      }
-      const comicImages = $("picture.item-comic-image img");
-      if (comicImages.length !== 1) {
-        return scrapeFailure(failureModes.GOCOMICS_MISSING_IMAGE_ON_PAGE);
-      }
-      return scrapeSuccess(comicImages.attr("src"));
-    }
-    case sites.comicskingdom.id: {
-      const url = `https://www.comicskingdom.com/${theirIdentifier}/${date.format(
-        "YYYY-MM-DD",
-      )}`;
-      const $ = await cheerioRequest(url);
-      if ($ === null) {
-        return scrapeFailure(failureModes.COMICS_KINGDOM_REJECTION);
-      }
-      const comicImages = $("meta[property='og:image']");
-      if (comicImages.length !== 1) {
-        return scrapeFailure(failureModes.COMICS_KINGDOM_MISSING_IMAGE_ON_PAGE);
-      }
-      return scrapeSuccess(comicImages.attr("content"));
-    }
-    case sites.arcamax.id:
-    case sites.arcamaxeditorials.id: {
-      const directory =
-        siteId === sites.arcamax.id
-          ? "thefunnies"
-          : "politics/editorialcartoons";
-      const url = `https://www.arcamax.com/${directory}/${theirIdentifier}/`;
-      const $ = await cheerioRequest(url);
-      if ($ === null) {
-        return scrapeFailure(failureModes.ARCAMAX_REJECTION);
-      }
-      const comicImages = $("img#comic-zoom");
-      if (comicImages.length !== 1) {
-        return scrapeFailure(failureModes.ARCAMAX_MISSING_IMAGE_ON_PAGE);
-      }
-      const imageUrl = comicImages.attr("src");
-      return scrapeSuccess(imageUrl);
-    }
-    case sites.tundra.id: {
-      const url = `http://www.tundracomicsftp.com/comicspub/daily_tundra/daily/${date.format(
-        "YYYY-M-D",
-      )}.jpg`;
-      if (await producesSuccessResponse(url)) {
-        return scrapeSuccess(url);
-      }
-      return scrapeFailure(failureModes.TUNDRA_IMAGE_404);
-    }
-    case sites.xkcd.id: {
-      const url = `https://xkcd.com/rss.xml`;
-      const $ = await cheerioRequest(url);
-      if ($ === null) {
-        return scrapeFailure(failureModes.XKCD_RSS_REJECTION);
-      }
-      const items = $("rss channel item");
-      if (items.length === 0) {
-        return scrapeFailure(failureModes.XKCD_PARSE_ERROR);
-      }
-      const firstItemDescriptions = items.first().find("description");
-      if (firstItemDescriptions.length === 0) {
-        return scrapeFailure(failureModes.XKCD_PARSE_ERROR);
-      }
-      const firstItemDescription = unescape(
-        firstItemDescriptions.first().html(),
-      );
-      const $$ = cheerio.load(firstItemDescription);
-      const comicImages = $$("img");
-      if (comicImages.length !== 1) {
-        return scrapeFailure(failureModes.XKCD_PARSE_ERROR);
-      }
-      return scrapeSuccess(comicImages.attr("src"), comicImages.attr("alt"));
-    }
-    // TODO(ecarrel): delete SMBC? It's unnecessary since we already get it from gocomics.
-    case sites.smbc.id: {
-      const url = `https://www.smbc-comics.com/comic/rss`;
-      const $ = await cheerioRequest(url);
-      if ($ === null) {
-        return scrapeFailure(failureModes.SMBC_RSS_REJECTION);
-      }
-      const items = $("rss channel item");
-      if (items.length === 0) {
-        return scrapeFailure(failureModes.SMBC_PARSE_ERROR);
-      }
-      const firstItemDescriptions = items.first().find("description");
-      if (firstItemDescriptions.length === 0) {
-        return scrapeFailure(failureModes.SMBC_PARSE_ERROR);
-      }
-      const firstItemDescription = unescape(
-        firstItemDescriptions.first().html(),
-      );
-      const $$ = cheerio.load(firstItemDescription);
-      const comicImages = $$("img");
-      if (comicImages.length !== 1) {
-        return scrapeFailure(failureModes.SMBC_PARSE_ERROR);
-      }
-      // TODO(ecarrel): scrape hover text too.
-      return scrapeSuccess(comicImages.attr("src"));
-    }
-    case sites.phdcomics.id: {
-      const url = `http://phdcomics.com/gradfeed.php`;
-      const $ = await cheerioRequest(url);
-      if ($ === null) {
-        return scrapeFailure(failureModes.PHD_COMICS_RSS_REJECTION);
-      }
-      const items = $("rss channel item");
-      if (items.length === 0) {
-        return scrapeFailure(failureModes.PHD_COMICS_PARSE_ERROR);
-      }
-      const firstItemDescriptions = items.first().find("description");
-      if (firstItemDescriptions.length === 0) {
-        return scrapeFailure(failureModes.PHD_COMICS_PARSE_ERROR);
-      }
-      const firstItemDescription = unescape(
-        firstItemDescriptions.first().html(),
-      );
-      const $$ = cheerio.load(firstItemDescription);
-      const comicImages = $$("img");
-      if (comicImages.length !== 1) {
-        return scrapeFailure(failureModes.PHD_COMICS_PARSE_ERROR);
-      }
-      // TODO(ecarrel): scrape hover text too.
-      return scrapeSuccess(comicImages.attr("src"));
-    }
-    case sites.explosm.id: {
-      const url = "http://explosm.net/comics/latest";
-      const $ = await cheerioRequest(url);
-      if ($ === null) {
-        return scrapeFailure(failureModes.EXPLOSM_REJECTION);
-      }
-      const comicImages = $("#comic img");
-      if (comicImages.length === 0) {
-        return scrapeFailure(failureModes.EXPLOSM_MISSING_IMAGE_ON_PAGE);
-      }
-      const imageUrl = comicImages.first().attr("src");
-      return scrapeSuccess(imageUrl);
-    }
-    case sites.dinosaur.id: {
-      const url = "http://www.qwantz.com/";
-      const $ = await cheerioRequest(url);
-      if ($ === null) {
-        return scrapeFailure(failureModes.DINOSAUR_REJECTION);
-      }
-      const comicImages = $("img.comic");
-      if (comicImages.length !== 1) {
-        return scrapeFailure(failureModes.DINOSAUR_MISSING_IMAGE_ON_PAGE);
-      }
-      const imageUrl = `${url}${comicImages.attr("src")}`;
-      return scrapeSuccess(imageUrl);
-    }
-    case sites.threewordphrase.id: {
-      const url = "http://threewordphrase.com/";
-      const $ = await cheerioRequest(url);
-      if ($ === null) {
-        return scrapeFailure(failureModes.THREE_WORD_PHRASE_REJECTION);
-      }
-      const comicImages = $("table tbody tr td center img");
-      if (comicImages.length !== 1) {
-        return scrapeFailure(
-          failureModes.THREE_WORD_PHRASE_MISSING_IMAGE_ON_PAGE,
-        );
-      }
-      const imageUrl = `${url}${comicImages.attr("src")}`;
-      return scrapeSuccess(imageUrl);
-    }
-    default: {
-      return scrapeFailure(failureModes.UNKNOWN_SITE);
-    }
+  const { site_id: siteId } = syndication;
+  const site = sites[siteId];
+  if (site == null) {
+    return scrapeFailure(failureModes.UNKNOWN_SITE);
   }
+  return site.scraper.scrape(date, syndication);
 };
 
 const createComicDbObject = (
