+++ api/src/service/scraper/common.ts
@@ -1,0 +1,76 @@
+import axios from "axios";
+import https from "https";
+import cheerio from "cheerio";
+import { ComicFailureMode, failureModes } from "../../db-models/comic-syndication";
+
+export const cheerioRequest = async (url: string) => {
+    try {
+        const response = await axios({
+            method: "GET",
+            url,
+            responseType: "text",
+            httpsAgent: new https.Agent({ rejectUnauthorized: false }),
+        });
+        // TODO(ecarrel): check if response.request.res.responseUrl !== url to detect redirects.
+        return cheerio.load(response.data);
+    } catch (err) {
+        return null;
+    }
+};
+
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
+// eslint-disable-next-line no-underscore-dangle,@typescript-eslint/no-unused-vars
+export const _producesNonEmptyResponse = async (url: string) => {
+    const response = await axios({
+        method: "GET",
+        url,
+        responseType: "stream",
+    }).catch(() => false);
+    if (typeof response === "boolean") {
+        return false;
+    }
+    const content = response.data.read(1);
+    return content != null;
+};
+
+export const producesSuccessResponse = async (url: string) => {
+    const response = await axios({
+        method: "GET",
+        url,
+        responseType: "stream",
+    }).catch(() => false);
+    if (typeof response === "boolean") {
+        return false;
+    }
+    return response.status === 200;
+};
+
+export type ScrapeResult = {
+    success: boolean;
+    imageUrl: string | null;
+    failureMode: ComicFailureMode | null;
+    imageCaption: string | null;
+};
+
+export const scrapeFailure = (failureMode: ComicFailureMode): ScrapeResult => ({
+    success: false,
+    imageUrl: null,
+    failureMode,
+    imageCaption: null,
+});
+
+export const scrapeSuccess = (
+    imageUrl: string | undefined,
+    imageCaption?: string | null,
+): ScrapeResult => {
+    if (imageUrl == null || imageUrl.length === 0) {
+        return scrapeFailure(failureModes.UNKNOWN);
+    }
+    return {
+        success: true,
+        imageUrl,
+        failureMode: null,
+        imageCaption: imageCaption || null,
+    };
+};
